encapsulates into calculator and report classes

+++ src/mortgageCalculator/Console.java
@@ -13,7 +13,7 @@ public class Console {
         double value;
         while (true) {
             System.out.print(prompt);
-            value = scanner.nextFloat();
+            value = scanner.nextDouble();
             if (value >= min && value <= max)
                 break;
             System.out.println("Enter a value between " + min + " and " + max);
